Refactor changed areas
