+++ apps/server/src/config/cors.ts
@@ -1,4 +1,4 @@
-import { env } from './env';
+import { env } from '../schemas/env';
 
 export const corsOptions = {
   origin:
