+++ apps/server/src/middleware/auth.ts
@@ -1,7 +1,7 @@
 import type { Request, Response, NextFunction } from 'express';
 import jwt from 'jsonwebtoken';
 
-import { env } from '../config/env';
+import { env } from '../schemas/env';
 
 const requireAuth = (req: Request, res: Response, next: NextFunction) => {
   try {
@@ -15,19 +15,25 @@ const requireAuth = (req: Request, res: Response, next: NextFunction) => {
 
     const decoded = jwt.verify(token, env.JWT_SECRET);
     if (typeof decoded === 'string') {
-      return res.status(400).json({ error: 'Invalid token format.' });
+      return res.status(401).json({ error: 'Invalid token format.' });
+    }
+
+    if (!decoded.userId || !decoded.username) {
+      return res.status(401).json({
+        error: 'Invalid token payload.',
+      });
     }
 
     // Set the user information in the request object for access in the controller
     req.user = {
-      userId: decoded.userId as number,
-      username: decoded.username as string,
+      userId: decoded.userId,
+      username: decoded.username,
       iat: decoded.iat!,
       exp: decoded.exp!,
     };
     next();
   } catch {
-    res.status(400).json({ error: 'Invalid token.' });
+    res.status(401).json({ error: 'Invalid token.' });
   }
 };
