+++ apps/server/src/schemas/env.ts
@@ -1,4 +1,5 @@
 import dotenv from 'dotenv';
+import { existsSync } from 'fs';
 import path from 'path';
 import { fileURLToPath } from 'url';
 import { z } from 'zod/v4';
@@ -6,7 +7,14 @@ import { z } from 'zod/v4';
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
-dotenv.config({ path: path.resolve(__dirname, '../../.env.local') });
+const envLocalPath = path.resolve(__dirname, '../../.env.local');
+const fallbackEnvPath = path.resolve(__dirname, '../../.env');
+
+if (existsSync(envLocalPath)) {
+  dotenv.config({ path: envLocalPath });
+} else {
+  dotenv.config({ path: fallbackEnvPath });
+}
 
 const envSchema = z.object({
   SUPABASE_URL: z.string(),
